Stonethorn update
Add keybind to add items in the mail to the queue
Added minimum version LLC dependency.
Now uses the instant multicrafting! If you have a multiplier, then it will only create one row, with a ##x label
Added a button to show/hide the favourites window, and removed the settings menu option
Favourites window will resize along with the main set crafter window
Fixed a bug where selecting a favourites would not select the correct armour weight
When sending a request, Jewelry items will no longer have N/A as the style, and it will instead be blank
Added a request to chat button, which will allow you to easily put your crafting request into chat. Also added the option to right click links to add to queue
Fixed some bugs with removing items from queue
Fixed a bug where the auto crafting option would not correctly be set on reloadui
Fixed a bug where after sustained use the dropdown tooltips would gradually stop working
Fixed a bug where the preview tooltips incorrectly had weapon enchants
Added an option to the favourites so that you can save the queue as a favourite
API Bump
Increased the minimum height of the window, so that when resizing the queue always has a backdrop